Attention cast floats

+++ perceiver_pytorch/perceiver_pytorch.py
@@ -14,13 +14,16 @@ from torch.nn import GELU
 def exists(val):
     return val is not None
 
+
 def default(val, d):
     return val if exists(val) else d
 
+
 def cache_fn(f):
     cache = None
+
     @wraps(f)
-    def cached_fn(*args, _cache = True, **kwargs):
+    def cached_fn(*args, _cache=True, **kwargs):
         if not _cache:
             return f(*args, **kwargs)
         nonlocal cache
@@ -28,24 +31,27 @@ def cache_fn(f):
             return cache
         cache = f(*args, **kwargs)
         return cache
+
     return cached_fn
 
-def fourier_encode(x, max_freq, num_bands = 4, base = 2):
+
+def fourier_encode(x, max_freq, num_bands=4, base=2):
     x = x.unsqueeze(-1)
     device, dtype, orig_x = x.device, x.dtype, x
 
-    scales = torch.logspace(1., log(max_freq / 2) / log(base), num_bands, base = base, device = device, dtype = dtype)
+    scales = torch.logspace(1., log(max_freq / 2) / log(base), num_bands, base=base, device=device, dtype=dtype)
     scales = scales[(*((None,) * (len(x.shape) - 1)), Ellipsis)]
 
     x = x * scales * pi
     x = torch.cat([x.sin(), x.cos()], dim=-1)
-    x = torch.cat((x, orig_x), dim = -1)
+    x = torch.cat((x, orig_x), dim=-1)
     return x
+
 
 # helper classes
 
 class PreNorm(nn.Module):
-    def __init__(self, dim, fn, context_dim = None):
+    def __init__(self, dim, fn, context_dim=None):
         super().__init__()
         self.fn = fn
         self.norm = nn.LayerNorm(dim)
@@ -57,17 +63,19 @@ class PreNorm(nn.Module):
         if exists(self.norm_context):
             context = kwargs['context']
             normed_context = self.norm_context(context)
-            kwargs.update(context = normed_context)
+            kwargs.update(context=normed_context)
 
         return self.fn(x, **kwargs)
 
+
 class GEGLU(nn.Module):
     def forward(self, x):
-        x, gates = x.chunk(2, dim = -1)
+        x, gates = x.chunk(2, dim=-1)
         return x * F.gelu(gates)
 
+
 class FeedForward(nn.Module):
-    def __init__(self, dim, mult = 4, dropout = 0.):
+    def __init__(self, dim, mult=4, dropout=0.):
         super().__init__()
         self.net = nn.Sequential(
             nn.Linear(dim, dim * mult * 2),
@@ -79,8 +87,9 @@ class FeedForward(nn.Module):
     def forward(self, x):
         return self.net(x)
 
+
 class FeedForwardGELU(nn.Module):
-    def __init__(self, dim, mult = 4, dropout = 0.):
+    def __init__(self, dim, mult=4, dropout=0.):
         super().__init__()
         self.net = nn.Sequential(
             nn.Linear(dim, dim * mult),
@@ -91,8 +100,10 @@ class FeedForwardGELU(nn.Module):
 
     def forward(self, x):
         return self.net(x)
+
+
 class Attention(nn.Module):
-    def __init__(self, query_dim, context_dim = None, heads = 8, dim_head = 64, dropout = 0.):
+    def __init__(self, query_dim, context_dim=None, heads=8, dim_head=64, dropout=0.):
         super().__init__()
         inner_dim = dim_head * heads
         context_dim = default(context_dim, query_dim)
@@ -100,60 +111,67 @@ class Attention(nn.Module):
         self.scale = dim_head ** -0.5
         self.heads = heads
 
-        self.to_q = nn.Linear(query_dim, inner_dim, bias = False)
-        self.to_kv = nn.Linear(context_dim, inner_dim * 2, bias = False)
+        self.to_q = nn.Linear(query_dim, inner_dim, bias=False)
+        self.to_kv = nn.Linear(context_dim, inner_dim * 2, bias=False)
 
         self.to_out = nn.Sequential(
             nn.Linear(inner_dim, query_dim),
             nn.Dropout(dropout)
         )
 
-    def forward(self, x, context = None, mask = None):
+    def forward(self, x, context=None, mask=None):
         h = self.heads
 
         q = self.to_q(x)
         context = default(context, x)
-        k, v = self.to_kv(context).chunk(2, dim = -1)
+        k, v = self.to_kv(context).chunk(2, dim=-1)
+        # Cast query and keys to float 32 to avoid instability as attention weights grow
+        # during training, per https://twitter.com/tsuname/status/1430653484827697155?s=20
+        k = k.float()
+        q = q.float()
 
-        q, k, v = map(lambda t: rearrange(t, 'b n (h d) -> (b h) n d', h = h), (q, k, v))
+        q, k, v = map(lambda t: rearrange(t, 'b n (h d) -> (b h) n d', h=h), (q, k, v))
 
         sim = einsum('b i d, b j d -> b i j', q, k) * self.scale
 
         if exists(mask):
             mask = rearrange(mask, 'b ... -> b (...)')
             max_neg_value = -torch.finfo(sim.dtype).max
-            mask = repeat(mask, 'b j -> (b h) () j', h = h)
+            mask = repeat(mask, 'b j -> (b h) () j', h=h)
             sim.masked_fill_(~mask, max_neg_value)
 
         # attention, what we cannot get enough of
-        attn = sim.softmax(dim = -1)
+        attn = sim.softmax(dim=-1)
 
         out = einsum('b i j, b j d -> b i d', attn, v)
-        out = rearrange(out, '(b h) n d -> b n (h d)', h = h)
+        # cast back to input type:
+        out = out.type(x.dtype)
+        out = rearrange(out, '(b h) n d -> b n (h d)', h=h)
         return self.to_out(out)
+
 
 # main class
 
 class Perceiver(nn.Module):
     def __init__(
-        self,
-        *,
-        num_freq_bands,
-        depth,
-        max_freq,
-        freq_base = 2,
-        input_channels = 3,
-        input_axis = 2,
-        num_latents = 512,
-        latent_dim = 512,
-        cross_heads = 1,
-        latent_heads = 8,
-        cross_dim_head = 64,
-        latent_dim_head = 64,
-        num_classes = 1000,
-        attn_dropout = 0.,
-        ff_dropout = 0.,
-        weight_tie_layers = False
+            self,
+            *,
+            num_freq_bands,
+            depth,
+            max_freq,
+            freq_base=2,
+            input_channels=3,
+            input_axis=2,
+            num_latents=512,
+            latent_dim=512,
+            cross_heads=1,
+            latent_heads=8,
+            cross_dim_head=64,
+            latent_dim_head=64,
+            num_classes=1000,
+            attn_dropout=0.,
+            ff_dropout=0.,
+            weight_tie_layers=False
     ):
         super().__init__()
         self.input_axis = input_axis
@@ -165,12 +183,17 @@ class Perceiver(nn.Module):
 
         self.latents = nn.Parameter(torch.randn(num_latents, latent_dim))
 
-        get_cross_attn = lambda: PreNorm(latent_dim, Attention(latent_dim, input_dim, heads = cross_heads, dim_head = cross_dim_head, dropout = attn_dropout), context_dim = input_dim)
-        get_cross_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout = ff_dropout))
-        get_latent_attn = lambda: PreNorm(latent_dim, Attention(latent_dim, heads = latent_heads, dim_head = latent_dim_head, dropout = attn_dropout))
-        get_latent_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout = ff_dropout))
+        get_cross_attn = lambda: PreNorm(latent_dim,
+                                         Attention(latent_dim, input_dim, heads=cross_heads, dim_head=cross_dim_head,
+                                                   dropout=attn_dropout), context_dim=input_dim)
+        get_cross_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout=ff_dropout))
+        get_latent_attn = lambda: PreNorm(latent_dim,
+                                          Attention(latent_dim, heads=latent_heads, dim_head=latent_dim_head,
+                                                    dropout=attn_dropout))
+        get_latent_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout=ff_dropout))
 
-        get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff = map(cache_fn, (get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff))
+        get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff = map(cache_fn, (
+        get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff))
 
         self.layers = nn.ModuleList([])
         for i in range(depth):
@@ -189,30 +212,30 @@ class Perceiver(nn.Module):
             nn.Linear(latent_dim, num_classes)
         )
 
-    def forward(self, data, mask = None):
+    def forward(self, data, mask=None):
         b, *axis, _, device = *data.shape, data.device
         assert len(axis) == self.input_axis, 'input data must have the right number of axis'
 
         # calculate fourier encoded positions in the range of [-1, 1], for all axis
 
-        axis_pos = list(map(lambda size: torch.linspace(-1., 1., steps = size, device = device), axis))
-        pos = torch.stack(torch.meshgrid(*axis_pos), dim = -1)
-        enc_pos = fourier_encode(pos, self.max_freq, self.num_freq_bands, base = self.freq_base)
+        axis_pos = list(map(lambda size: torch.linspace(-1., 1., steps=size, device=device), axis))
+        pos = torch.stack(torch.meshgrid(*axis_pos), dim=-1)
+        enc_pos = fourier_encode(pos, self.max_freq, self.num_freq_bands, base=self.freq_base)
         enc_pos = rearrange(enc_pos, '... n d -> ... (n d)')
-        enc_pos = repeat(enc_pos, '... -> b ...', b = b)
+        enc_pos = repeat(enc_pos, '... -> b ...', b=b)
 
         # concat to channels of data and flatten axis
 
-        data = torch.cat((data, enc_pos), dim = -1)
+        data = torch.cat((data, enc_pos), dim=-1)
         data = rearrange(data, 'b ... d -> b (...) d')
 
-        x = repeat(self.latents, 'n d -> b n d', b = b)
+        x = repeat(self.latents, 'n d -> b n d', b=b)
 
         for cross_attn, cross_ff, latent_attn, latent_ff in self.layers:
-            x = cross_attn(x, context = data, mask = mask) + x
+            x = cross_attn(x, context=data, mask=mask) + x
             x = cross_ff(x) + x
             x = latent_attn(x) + x
             x = latent_ff(x) + x
 
-        x = x.mean(dim = -2)
+        x = x.mean(dim=-2)
         return self.to_logits(x)
